+++ next.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+	async rewrites() {
+		return [
+			{
+				source: '/:any*',
+				destination: '/',
+			},
+		];
+	},
+};
