+++ package.json
@@ -35,6 +35,7 @@
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^5.3.0",
     "react-side-effect": "^2.1.1",
     "shallow-equal": "^1.2.1",
     "shallowequal": "^1.1.0",
