+++ src/pages/_app.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { AppProps } from 'next/app';
 import '../../styles/globals.css';
 
-function MemoDraw({Component, pageProps} : AppProps): JSX.Element {
-	return <Component {...pageProps} />
+function MemoDraw({ Component, pageProps }: AppProps): JSX.Element {
+	return typeof window === 'undefined' ? null : <Component {...pageProps} />
 }
 
 export default MemoDraw;
